Adjust for different bases of lines and columns
The used CodeBlocks methods use 0-based numbers but the protocol requires 1-based numbers.

TODO: The column handling is kind of weird in SourceTrail. It is possible to select column 0 which leads in CodeBlocks to move one line up. A Click on an character looks like in CodeBlocks one characters too far left gets selected. But without the adjustment a message from CodeBlocks would be one too far left and a token on the cursor won't be selected.